refactor(auto): share built-in auto catalog via a module; rebaseline file-size
Review adjustment on top of the built-in auto/* catalog fix: the
AUTO_TEMPLATE_VARIANTS / VALID_AUTO_VARIANTS maps were duplicated inline in
both chat.ts (handleChat) and chatHelpers.ts (resolveModelOrError). Extract
them — plus resolveAutoVariant/createBuiltinAutoCombo — into
open-sse/services/autoCombo/builtinCatalog.ts (60 LOC), imported by both
handlers. chatHelpers.ts drops back under the 800-LOC cap; chat.ts keeps only
its recognition-gating logic and its file-size baseline is bumped 1432->1458
(genuine new logic, data block now shared). Behavior unchanged.
